Added Enemy floor collision / Improved Cutscenes & Animations
Enemies will now be able to handle tile collisions.
Improved the cutscene code (More generalized + no jittering)
Improved Shade animations to look more smooth

+++ sprite.h
@@ -24,7 +24,7 @@ public:
 	virtual ~Sprite();
 	virtual void update();
 	virtual void updateBullet(int bX, int bY, int bW, int bH);
-	virtual void updateBoss(int y);
+	virtual void updateBoss(float y);
 	const Rectangle getProjectileBBox() const;
 	void draw(Graphics &graphics, int x, int y);
 	void drawHPBar(Graphics &graphics, int x, int y);
